pyproject.toml
[virtualenvs]
create = true
in-project = true
prefer-active-python = true

[tool.poetry]
name = "indico-patcher"
version = "0.2.1"
description = "The Swiss Army knife to customize Indico"
repository = "https://github.com/unconventionaldotdev/indico-patcher"
readme = "README.md"
license = "MIT"
authors = [
    "Alejandro Avilés <ome@unconventional.dev>"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Application Frameworks"
]

[tool.poetry.dependencies]
python = ">=3.12.2,<3.13"
aenum = ">=3.1.15"
indico = ">=3.3"

[tool.poetry.group.dev.dependencies]
mypy = "^1.6.1"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
ruff = "^0.3.5"
sqlalchemy2-stubs = "^0.0.2a35"
tox = "^4.11.4"
unbeheader = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
